연 뭉게구름 완료 및 장애 해결
- 연출 외부 제어 가능
- 밀기와 대미지 로직을 합친 마커 구현 및 테스트 완료

+++ Assets/GamePlay/Features/Battle/Scripts/BattleTurn/Turn.cs
@@ -115,7 +115,7 @@ namespace GamePlay.Features.Battle.Scripts.BattleTurn
         /// <summary>
         /// 행동 수행 가능 여부 검증
         /// </summary>
-        public bool CanPerformAction(TurnActionState.ActionCategory category, float moveDistance = 0f)
+        public bool CanPerformAction(TurnActionState.ActionCategory category, int moveDistance=0)
         {
             return ActionState.CanPerformAction(category, moveDistance);
         }
